Fix "InvalidOperationException: Trying to build visual line from collapsed line":
When the last line in the document was collapsed, the method HeightTree.GetNodeByVisualPosition(position after end of document)
was returning the last line in the document, when it should have returned the last non-collapsed line.

Also fixed a minor bug in FoldingMargin (missing end marker when folding starts above viewport and ends on last line of document).